+++ NEWS.md
@@ -1,3 +1,7 @@
-# coefa 1.0.1
+## **Changes in version 1.0.2**
 
-* Initial release.
+-   Added a function to fix the value of diagonal in similarity matrix
+
+-   Added the function of `coefa_acm` ,the raw co-occurrence matrix can be displayed
+
+-   Updated Introduction